Moved ScreenOverlayEnum to utilities package
