+++ core/src/com/mygdx/chalmersdefense/controllers/overlays/SettingsOverlayController.java
@@ -8,8 +8,7 @@ import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.mygdx.chalmersdefense.model.IControllModel;
 import com.mygdx.chalmersdefense.utilities.Preferences;
-import com.mygdx.chalmersdefense.views.overlays.OverlayManager;
-import com.mygdx.chalmersdefense.views.overlays.ScreenOverlayEnum;
+import com.mygdx.chalmersdefense.utilities.ScreenOverlayEnum;
 
 /**
  * @author Daniel Persson
