+++ core/src/com/mygdx/chalmersdefense/model/IViewModel.java
@@ -1,6 +1,6 @@
 package com.mygdx.chalmersdefense.model;
 
-import com.mygdx.chalmersdefense.views.overlays.ScreenOverlayEnum;
+import com.mygdx.chalmersdefense.utilities.ScreenOverlayEnum;
 import com.mygdx.chalmersdefense.model.targetMode.ITargetMode;
 import com.mygdx.chalmersdefense.model.viruses.IVirus;
 import com.mygdx.chalmersdefense.utilities.GetRangeCircle;
