+++ core/src/com/mygdx/chalmersdefense/views/overlays/OverlayManager.java
@@ -1,8 +1,8 @@
 package com.mygdx.chalmersdefense.views.overlays;
 
+import com.mygdx.chalmersdefense.utilities.ScreenOverlayEnum;
 import com.mygdx.chalmersdefense.views.AbstractScreen;
 import com.mygdx.chalmersdefense.views.GameScreenViews.*;
-import com.mygdx.chalmersdefense.views.ScreenEnum;
 import com.mygdx.chalmersdefense.views.ScreenManager;
 
 public class OverlayManager {
